Feat: 비밀번호 변경, 업로드시 디폴트이미지 파일 삭제 안되게 수정

+++ src/main/java/com/sparta/zipsa/service/file/FileServiceImpl.java
@@ -36,6 +36,8 @@ public class FileServiceImpl implements FileService {
     //파일 삭제
     @Override
     public void deleteFile(String filename) {
-        new File(location + filename).delete();
+        if(!filename.equals("920ba86-defaultImage.jpg")){
+            new File(location + filename).delete();
+        }
     }
 }
